wip - refactor rendering
